Revert "feat: sdk observables layer (#662)"
This reverts commit 7ff4169e4fbe9b9eeb02ec510448c01a81bb6ef1.
