+++ Explorer/Assets/DCL/Multiplayer/SDK/Systems/GlobalWorld/AvatarEmoteCommandPropagationSystem.cs
@@ -1,18 +1,19 @@
 using Arch.Core;
 using Arch.System;
 using Arch.SystemGroups;
-using Arch.SystemGroups.DefaultSystemGroups;
 using DCL.AvatarRendering.Emotes;
 using DCL.Diagnostics;
 using DCL.Multiplayer.SDK.Components;
 using ECS.Abstract;
+using ECS.Groups;
 using ECS.LifeCycle.Components;
 using SceneRunner.Scene;
 
 namespace DCL.Multiplayer.SDK.Systems.GlobalWorld
 {
-    [UpdateInGroup(typeof(PresentationSystemGroup))]
-    [UpdateBefore(typeof(CharacterEmoteSystem))]
+    [UpdateInGroup(typeof(SyncedPreRenderingSystemGroup))]
+    [UpdateAfter(typeof(PlayerCRDTEntitiesHandlerSystem))]
+    [UpdateBefore(typeof(CleanUpGroup))]
     [LogCategory(ReportCategory.MULTIPLAYER_SDK_EMOTE_COMMAND_DATA)]
     public partial class AvatarEmoteCommandPropagationSystem : BaseUnityLoopSystem
     {
@@ -30,7 +31,7 @@ namespace DCL.Multiplayer.SDK.Systems.GlobalWorld
 
         [Query]
         [None(typeof(DeleteEntityIntention))]
-        private void UpdateEmoteCommandDataComponent(PlayerCRDTEntity playerCRDTEntity, CharacterEmoteIntent emoteIntent)
+        private void UpdateEmoteCommandDataComponent(ref PlayerCRDTEntity playerCRDTEntity, ref CharacterEmoteIntent emoteIntent)
         {
             SceneEcsExecutor sceneEcsExecutor = playerCRDTEntity.SceneFacade.EcsExecutor;
             World sceneWorld = sceneEcsExecutor.World;
@@ -46,10 +47,13 @@ namespace DCL.Multiplayer.SDK.Systems.GlobalWorld
                 emoteCommandComponent.PlayingEmote = emoteIntent.EmoteId;
                 emoteCommandComponent.LoopingEmote = emote.IsLooping();
 
-                if (componentFound)
-                    sceneWorld.Set(playerCRDTEntity.SceneWorldEntity, emoteCommandComponent);
-                else
-                    sceneWorld.Add(playerCRDTEntity.SceneWorldEntity, emoteCommandComponent);
+                // External world access should be always synchronized (Global World calls into Scene World)
+                {
+                    if (componentFound)
+                        sceneWorld.Set(playerCRDTEntity.SceneWorldEntity, emoteCommandComponent);
+                    else
+                        sceneWorld.Add(playerCRDTEntity.SceneWorldEntity, emoteCommandComponent);
+                }
             }
         }
     }
