+++ Explorer/Assets/DCL/Multiplayer/SDK/Systems/SceneWorld/WriteAvatarEquippedDataSystem.cs
@@ -16,7 +16,6 @@ using ECS.LifeCycle.Systems;
 namespace DCL.Multiplayer.SDK.Systems.SceneWorld
 {
     [UpdateInGroup(typeof(SyncedPreRenderingSystemGroup))]
-    [UpdateAfter(typeof(WritePlayerIdentityDataSystem))]
     [UpdateBefore(typeof(ResetDirtyFlagSystem<Profile>))]
     [LogCategory(ReportCategory.PLAYER_AVATAR_EQUIPPED)]
     public partial class WriteAvatarEquippedDataSystem : BaseUnityLoopSystem
