+++ Explorer/Assets/Scripts/CrdtEcsBridge/CRDT.ECS.Bridge.asmdef
@@ -11,8 +11,7 @@
         "GUID:101b8b6ebaf64668909b49c4b7a1420d",
         "GUID:e0cd26848372d4e5c891c569017e11f1",
       "GUID:275e22790c04e9b47a5085d7b0c4432a",
-      "GUID:3640f3c0b42946b0b8794a1ed8e06ca5",
-      "GUID:b4018cfd60a549f6ab9f20b9a5db4a56"
+      "GUID:3640f3c0b42946b0b8794a1ed8e06ca5"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
