--- Explorer/Assets/Scripts/CrdtEcsBridge/JsModulesImplementation/Communications/CommunicationsControllerAPIImplementationBase.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: cd97f4e868344b1aaac96b5e89dd132f
-timeCreated: 1715588972
