--- Explorer/Assets/Scripts/CrdtEcsBridge/JsModulesImplementation/Communications/SDKMessageBus.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 3040fc1f242d4aa9b8c84f13869e4327
-timeCreated: 1714580745
