--- Explorer/Assets/Scripts/CrdtEcsBridge/JsModulesImplementation/Communications/SDKMessageBus/SDKMessageBusCommsAPIImplementation.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: fdf741d159e141469149fefc3469add2
-timeCreated: 1714580773
