+++ Explorer/Assets/Scripts/CrdtEcsBridge/JsModulesImplementation/JsModulesImplementation.asmdef
@@ -24,8 +24,7 @@
         "GUID:3c7b57a14671040bd8c549056adc04f5",
         "GUID:875a5d5129614170bd769ed012c2eb3d",
         "GUID:702f733b4deb246808c6ce84d93b5c9c",
-        "GUID:166b65e6dfc848bb9fb075f53c293a38",
-        "GUID:b4018cfd60a549f6ab9f20b9a5db4a56"
+        "GUID:166b65e6dfc848bb9fb075f53c293a38"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
