--- Explorer/Assets/Scripts/CrdtEcsBridge/JsModulesImplementation/SDKObservableEventsEngineAPIImplementation.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: a7ca9369bac5468e9b4a6d1356ef41b3
-timeCreated: 1714676966
