+++ Explorer/Assets/Scripts/CrdtEcsBridge/JsModulesImplementation/Tests/EngineAPIImplementationShould.cs
@@ -109,7 +109,7 @@ namespace CrdtEcsBridge.JsModulesImplementation.Tests
             sharedPoolsProvider.GetSerializedStateBytesPool(Arg.Any<int>()).Returns(c => new PoolableByteArray(new byte[c.Arg<int>()], c.Arg<int>(), sharedPoolsProvider.ReleaseSerializedStateBytesPool));
             sharedPoolsProvider.GetSerializationCrdtMessagesPool(Arg.Any<int>()).Returns(c => new ProcessedCRDTMessage[c.Arg<int>()]);
 
-            outgoingCrtdMessagesProvider.GetSerializationSyncBlock(Arg.Any<Action<OutgoingCRDTMessagesProvider.PendingMessage>>())
+            outgoingCrtdMessagesProvider.GetSerializationSyncBlock()
                                         .Returns(_ =>
                                          {
                                              mutex.WaitOne();
@@ -146,7 +146,7 @@ namespace CrdtEcsBridge.JsModulesImplementation.Tests
                 worldSyncCommandBuffer.FinalizeAndDeserialize();
                 instancePoolsProvider.ReleaseDeserializationMessagesPool(Arg.Any<List<CRDTMessage>>());
 
-                outgoingCrtdMessagesProvider.GetSerializationSyncBlock(Arg.Any<Action<OutgoingCRDTMessagesProvider.PendingMessage>>());
+                outgoingCrtdMessagesProvider.GetSerializationSyncBlock();
                 sharedPoolsProvider.GetSerializedStateBytesPool(Arg.Any<int>());
 
                 // can't check Serialize, can't mock `Span<byte>`
@@ -162,7 +162,7 @@ namespace CrdtEcsBridge.JsModulesImplementation.Tests
 
             Received.InOrder(() =>
             {
-                outgoingCrtdMessagesProvider.GetSerializationSyncBlock(Arg.Any<Action<OutgoingCRDTMessagesProvider.PendingMessage>>());
+                outgoingCrtdMessagesProvider.GetSerializationSyncBlock();
                 crdtProtocol.GetMessagesCount();
                 sharedPoolsProvider.GetSerializationCrdtMessagesPool(crdtStateMessages.Count + outgoingMessages.Count);
                 crdtProtocol.CreateMessagesFromTheCurrentState(Arg.Any<ProcessedCRDTMessage[]>());
