+++ Explorer/Assets/Scripts/CrdtEcsBridge/OutgoingMessages/IOutgoingCRDTMessagesProvider.cs
@@ -30,6 +30,6 @@ namespace CrdtEcsBridge.OutgoingMessages
         ///     Freeze the modification of the queue while it's being processed from the background thread
         /// </summary>
         /// <returns></returns>
-        OutgoingCRDTMessagesSyncBlock GetSerializationSyncBlock([CanBeNull] Action<OutgoingCRDTMessagesProvider.PendingMessage> actOnPendingMessage);
+        OutgoingCRDTMessagesSyncBlock GetSerializationSyncBlock();
     }
 }
