+++ Explorer/Assets/Scripts/CrdtEcsBridge/PoolsProviders/ISharedPoolsProvider.cs
@@ -1,5 +1,4 @@
 using CRDT.Protocol.Factory;
-using SceneRuntime.Apis.Modules.EngineApi.SDKObservableEvents.Events;
 
 namespace CrdtEcsBridge.PoolsProviders
 {
@@ -13,12 +12,8 @@ namespace CrdtEcsBridge.PoolsProviders
 
         PoolableByteArray GetSerializedStateBytesPool(int size);
 
-        PoolableSDKObservableEventArray GetSerializationSDKObservableEventsPool(int size);
-
         void ReleaseSerializationCrdtMessagesPool(ProcessedCRDTMessage[] messages);
 
         void ReleaseSerializedStateBytesPool(byte[] bytes);
-
-        void ReleaseSerializationSDKObservableEventsPool(SDKObservableEvent[] events);
     }
 }
