--- Explorer/Assets/Scripts/CrdtEcsBridge/PoolsProviders/PoolableSDKObservableEventArray.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: ee0af02e3b854fba841e975a61353332
-timeCreated: 1715703349
