+++ Explorer/Assets/Scripts/CrdtEcsBridge/PoolsProviders/SharedPoolsProvider.cs
@@ -1,5 +1,4 @@
 using CRDT.Protocol.Factory;
-using SceneRuntime.Apis.Modules.EngineApi.SDKObservableEvents.Events;
 using System;
 using System.Buffers;
 
@@ -15,15 +14,11 @@ namespace CrdtEcsBridge.PoolsProviders
         // Must be synchronized
         private static readonly ArrayPool<byte> SERIALIZED_STATE_BYTES_POOL = ArrayPool<byte>.Create();
 
-        private static readonly ArrayPool<SDKObservableEvent> SDK_OBSERVABLE_EVENTS_POOL = ArrayPool<SDKObservableEvent>.Create();
-
-        private readonly Action<byte[]> bytesPoolReleaseFuncCached;
-        private readonly Action<SDKObservableEvent[]> sdkObservableEventsPoolReleaseFuncCached;
+        private readonly Action<byte[]> releaseFuncCached;
 
         public SharedPoolsProvider()
         {
-            bytesPoolReleaseFuncCached = ReleaseSerializedStateBytesPool;
-            sdkObservableEventsPoolReleaseFuncCached = ReleaseSerializationSDKObservableEventsPool;
+            releaseFuncCached = ReleaseSerializedStateBytesPool;
         }
 
         public ProcessedCRDTMessage[] GetSerializationCrdtMessagesPool(int size)
@@ -38,22 +33,12 @@ namespace CrdtEcsBridge.PoolsProviders
 
         public PoolableByteArray GetSerializedStateBytesPool(int size)
         {
-            lock (SERIALIZED_STATE_BYTES_POOL) { return new PoolableByteArray(SERIALIZED_STATE_BYTES_POOL.Rent(size), size, bytesPoolReleaseFuncCached); }
+            lock (SERIALIZED_STATE_BYTES_POOL) { return new PoolableByteArray(SERIALIZED_STATE_BYTES_POOL.Rent(size), size, releaseFuncCached); }
         }
 
         public void ReleaseSerializedStateBytesPool(byte[] bytes)
         {
             lock (SERIALIZED_STATE_BYTES_POOL) { SERIALIZED_STATE_BYTES_POOL.Return(bytes, true); }
-        }
-
-        public PoolableSDKObservableEventArray GetSerializationSDKObservableEventsPool(int size)
-        {
-            lock (SDK_OBSERVABLE_EVENTS_POOL) { return new PoolableSDKObservableEventArray(SDK_OBSERVABLE_EVENTS_POOL.Rent(size), size, sdkObservableEventsPoolReleaseFuncCached); }
-        }
-
-        public void ReleaseSerializationSDKObservableEventsPool(SDKObservableEvent[] events)
-        {
-            lock (SDK_OBSERVABLE_EVENTS_POOL) { SDK_OBSERVABLE_EVENTS_POOL.Return(events, true); }
         }
     }
 }
