--- Explorer/Assets/Scripts/SceneRuntime/Apis/Modules/CommunicationsControllerApi/SDKMessageBus.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 5b7266205d6c4132b61fe9863af78ee3
-timeCreated: 1714566147
