--- Explorer/Assets/Scripts/SceneRuntime/Apis/Modules/CommunicationsControllerApi/SDKMessageBus/ISDKMessageBusCommsControllerAPI.cs
@@ -1,12 +1,0 @@
-using SceneRuntime.Apis.Modules.EngineApi.SDKObservableEvents.Events;
-using System.Collections.Generic;
-
-namespace SceneRuntime.Apis.Modules.CommunicationsControllerApi.SDKMessageBus
-{
-    public interface ISDKMessageBusCommsControllerAPI : ICommunicationsControllerAPI
-    {
-        List<CommsPayload> SceneCommsMessages { get; }
-        void Send(string data);
-    }
-}
-
