--- Explorer/Assets/Scripts/SceneRuntime/Apis/Modules/CommunicationsControllerApi/SDKMessageBus/ISDKMessageBusCommsControllerAPI.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 92e398c57ada4c8699ec7079ed365ca3
-timeCreated: 1714583439
