--- Explorer/Assets/Scripts/SceneRuntime/Apis/Modules/CommunicationsControllerApi/SDKMessageBus/SDKMessageBusCommsControllerAPIWrapper.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 4deb3e387ca849779fac42310f6c4a20
-timeCreated: 1714515845
