+++ Explorer/Assets/Scripts/SceneRuntime/Apis/Modules/EngineApi/EngineApiWrapper.cs
@@ -2,7 +2,6 @@
 using JetBrains.Annotations;
 using Microsoft.ClearScript.JavaScript;
 using SceneRunner.Scene.ExceptionsHandling;
-using SceneRuntime.Apis.Modules.EngineApi.SDKObservableEvents;
 using System;
 using UnityEngine.Profiling;
 
@@ -13,7 +12,7 @@ namespace SceneRuntime.Apis.Modules.EngineApi
         internal readonly IEngineApi api;
 
         private readonly IInstancePoolsProvider instancePoolsProvider;
-        protected readonly ISceneExceptionsHandler exceptionsHandler;
+        private readonly ISceneExceptionsHandler exceptionsHandler;
 
         private PoolableByteArray lastInput = PoolableByteArray.EMPTY;
 
@@ -72,9 +71,6 @@ namespace SceneRuntime.Apis.Modules.EngineApi
                 return ScriptableByteArray.EMPTY;
             }
         }
-
-        [UsedImplicitly]
-        public virtual ScriptableSDKObservableEventArray SendBatch() => null;
 
         public void SetIsDisposing()
         {
