--- Explorer/Assets/Scripts/SceneRuntime/Apis/Modules/EngineApi/SDKObservableEvents.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9902ce65fb1da4954b2fe8e97d77ab16
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
