--- Explorer/Assets/Scripts/SceneRuntime/Apis/Modules/EngineApi/SDKObservableEvents/Events.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 5c1ffbfe2d2e42f094d66714897c1a6a
-timeCreated: 1714582353
