--- Explorer/Assets/Scripts/SceneRuntime/Apis/Modules/EngineApi/SDKObservableEvents/Events/SDKObservableEvents.asmdef
@@ -1,15 +1,0 @@
-{
-    "name": "SDKObservableEvents",
-    "rootNamespace": "",
-    "references": [
-    "GUID:3c7b57a14671040bd8c549056adc04f5"],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
