--- Explorer/Assets/Scripts/SceneRuntime/Apis/Modules/EngineApi/SDKObservableEvents/Events/SDKObservableEvents.asmdef.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: b4018cfd60a549f6ab9f20b9a5db4a56
-timeCreated: 1714582408
