--- Explorer/Assets/Scripts/SceneRuntime/Apis/Modules/EngineApi/SDKObservableEvents/ISDKObservableEventsEngineApi.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 00bf3bd98f7840348c7380bd75f4597c
-timeCreated: 1714677333
