--- Explorer/Assets/Scripts/SceneRuntime/Apis/Modules/EngineApi/SDKObservableEvents/ScriptableSDKObservableEventArray.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: bc912f5ba49f4390884b7f5a78ccf132
-timeCreated: 1715704065
