+++ Explorer/Assets/Scripts/SceneRuntime/SceneRuntime.asmdef
@@ -17,10 +17,7 @@
         "GUID:166b65e6dfc848bb9fb075f53c293a38",
         "GUID:e25ef972de004615a22937e739de2def",
         "GUID:4a12c0b1b77ec6b418a8d7bd5c925be3",
-        "GUID:e0eedfa2deb9406daf86fd8368728e39",
-        "GUID:0f4c0f120707fb74497f5d581b9858f8",
-        "GUID:3c7b57a14671040bd8c549056adc04f5",
-        "GUID:b4018cfd60a549f6ab9f20b9a5db4a56"
+        "GUID:e0eedfa2deb9406daf86fd8368728e39"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
